+++ SIG_Dentistry.c
@@ -1,30 +1,21 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <locale.h>
+#include "funcoes_paciente.h"
+#include "funcoes_servico.h"
+#include "funcoes_funcionarios.h"
 
 //Assinatura das funcoes
 void tela_progr_odont();
 
 //Funções pacientes
 void tela_pacientes();
-void tela_cadastrar_paciente();
-void tela_pesquisar_paciente();
-void tela_alterar_paciente();
-void tela_excluir_paciente();
 
 //Funções serviços
 void tela_servicos();
-void tela_cadastrar_servico();
-void tela_pesquisar_servico();
-void tela_alterar_servico();
-void tela_excluir_servico();
 
 //Funções funcionários
 void tela_funcionarios();
-void tela_cadastrar_funcionario();
-void tela_pesquisar_funcionario();
-void tela_alterar_funcionario();
-void tela_excluir_funcionario();
 
 //Funções Agendamento
 void tela_agend();
@@ -82,7 +73,6 @@ void tela_progr_odont(){
         }             
     }while(opcao_navegacao!='0');    
     printf("\tSaindo do programa...\n");
-    sleep(1);
     system("cls||clear");
 }
 
@@ -125,94 +115,6 @@ void tela_pacientes(){
     }while(voltar);    
 }
 
-void tela_cadastrar_paciente(){
-    char cpf_paciente[14];
-    char nome_paciente[51];
-    char idade_paciente[3];
-    char endereco_paciente[51];
-    char email_paciente[51];
-    char numero_paciente[12];
-    do{
-        system ("cls||clear");
-        printf("\t===================================================\n");
-        printf("\t===============   Cadastrar Paciente   ============\n");
-        printf("\t===================================================\n\n");
-        printf("\t === Insira o CPF:    ");
-        scanf("%[0-9 .-/]", cpf_paciente);
-        getchar();
-        printf("\t === Insira o nome:   ");
-        scanf("%[A-ZÁÉÍÓÚÂÊÔÇÀÃÕ a-záéíóúâêôçàãõ]", nome_paciente);
-        getchar();
-        printf("\t === Insira a idade:  ");
-        scanf("%s", idade_paciente);
-        getchar();
-        printf("\t === Insira o endereco:   ");
-        scanf("%[A-ZÁÉÍÓÚÂÊÔÇÀÃÕ a-záéíóúâêôçàãõ 0-9 -- .]", endereco_paciente);
-        getchar();
-        printf("\t === Insira o e-mail: ");
-        scanf("%[A-Za-z0-9@._]", email_paciente);
-        getchar();
-        printf("\t === Insira o numero telefonico:  ");
-        scanf("%[0-9]", numero_paciente);
-        getchar();
-        printf("\t==================================================\n\n");
-        printf("\t === Digite 0 para confirmar\n");
-        printf("\t=>");
-        scanf("%i", &voltar);
-        getchar();
-    }while(voltar);         
-}
-
-void tela_pesquisar_paciente(){
-    char cpf_paciente[14];
-    do{
-        system ("cls||clear");
-        printf("\t===================================================\n");
-        printf("\t===============   Pesquisar Paciente   ============\n");
-        printf("\t===================================================\n\n");
-        printf("\t === Digite o CPF:    ");
-        scanf("%[0-9 .-/]", cpf_paciente);
-        printf("\t==================================================\n\n");
-        printf("\t === Digite 0 para voltar\n");
-        printf("\t=>");
-        scanf("%i", &voltar);
-        getchar();
-    }while(voltar);
-}
-
-void tela_alterar_paciente(){
-    char cpf_paciente[14];
-    do{
-        system ("cls||clear");
-        printf("\t===================================================\n");
-        printf("\t================   Alterar Paciente   =============\n");
-        printf("\t===================================================\n\n");
-        printf("\t === Digite o CPF:    ");
-        scanf("%[0-9 .-/]", cpf_paciente);
-        printf("\t==================================================\n\n");
-        printf("\t === Digite 0 para confirmar\n");
-        printf("\t=>");
-        scanf("%i", &voltar);
-        getchar();
-    }while(voltar);
-}
-
-void tela_excluir_paciente(){
-    char cpf_paciente[14];
-    do{
-        system ("cls||clear");
-        printf("\t===================================================\n");
-        printf("\t==============   Pesquisar Paciente   =============\n");
-        printf("\t===================================================\n\n");
-        printf("\t === Digite o CPF:    ");
-        scanf("%[0-9 .-/]", cpf_paciente);
-        printf("\t==================================================\n\n");
-        printf("\t=>");
-        scanf("%i", &voltar);
-        getchar();
-    }while(voltar);
-}
-
 void tela_servicos()
 {  
     do
@@ -253,69 +155,6 @@ void tela_servicos()
     }while(voltar); 
 }
 
-void tela_cadastrar_servico(){
-    do{
-        system ("cls||clear");
-        printf("\t===================================================\n");
-        printf("\t==============   Cadastrar Servicos   =============\n");
-        printf("\t===================================================\n\n");
-        printf("\t === Digite o nome do servico:\n");
-        printf("\t === Digite o custo do servico:\n");
-        printf("\t === Digite o tempo do procedimento:\n");
-        printf("\t === Digite a disponibilidade:\n");
-        printf("\t==================================================\n\n");
-        printf("\t === Digite 0 para confirmar\n");
-        printf("\t=>");
-        scanf("%i", &voltar);
-        getchar();
-    }while(voltar);         
-}
-
-void tela_pesquisar_servico(){
-    do{
-        system ("cls||clear");
-        printf("\t===================================================\n");
-        printf("\t==============   Pesquisar Servicos   =============\n");
-        printf("\t===================================================\n\n");
-        printf("\t === EM DESENVOLVIMENTO... \n");
-        printf("\t==================================================\n\n");
-        printf("\t === Digite 0 para voltar\n");
-        printf("\t=>");
-        scanf("%i", &voltar);
-        getchar();
-    }while(voltar);
-}
-
-void tela_alterar_servico(){
-    do{
-        system ("cls||clear");
-        printf("\t===================================================\n");
-        printf("\t===============   Alterar Servicos   ==============\n");
-        printf("\t===================================================\n\n");
-        printf("\t ===  EM DESENVOLVIMENTO... \n");
-        printf("\t==================================================\n\n");
-        printf("\t === Digite 0 para confirmar\n");
-        printf("\t=>");
-        scanf("%i", &voltar);
-        getchar();
-    }while(voltar);
-}
-
-void tela_excluir_servico(){
-    do{
-        system ("cls||clear");
-        printf("\t===================================================\n");
-        printf("\t===============   Excluir Servicos   ==============\n");
-        printf("\t===================================================\n\n");
-        printf("\t === EM DESENVOLVIMENTO... \n");
-        printf("\t==================================================\n\n");
-        printf("\t === Digite 0 para confirmar a exclusao\n");
-        printf("\t=>");
-        scanf("%i", &voltar);
-        getchar();
-    }while(voltar);
-}
-
 void tela_funcionarios(){  
     do
     {
@@ -353,95 +192,6 @@ void tela_funcionarios(){
 
         }
     }while(voltar);  
-}
-
-void tela_cadastrar_funcionario(){
-    char cpf_funcionario[14];
-    char nome_funcionario[51];
-    char idade_funcionario[3];
-    char endereco_funcionario[51];
-    char email_funcionario[51];
-    char numero_funcionario[12];
-    do{
-        system ("cls||clear");
-        printf("\t=========================================================\n");
-        printf("\t===============   Cadastrar Funcionarios   ==============\n");
-        printf("\t=========================================================\n\n");
-        printf("\t === Insira o CPF:    ");
-        scanf("%[0-9./]", cpf_funcionario);
-        getchar();
-        printf("\t === Insira o nome:   ");
-        scanf("%[A-ZÁÉÍÓÚÂÊÔÇÀÃÕ a-záéíóúâêôçàãõ]", nome_funcionario);
-        getchar();
-        printf("\t === Insira a idade:  ");
-        scanf("%s", idade_funcionario);
-        getchar();
-        printf("\t === Insira o endereco:   ");
-        scanf("%[A-ZÁÉÍÓÚÂÊÔÇÀÃÕ a-záéíóúâêôçàãõ 0-9 -- .]", endereco_funcionario);
-        getchar();
-        printf("\t === Insira o e-mail: ");
-        scanf("%[A-Za-z0-9@._]", email_funcionario);
-        getchar();
-        printf("\t === Insira o numero telefonico:  ");
-        scanf("%[0-9]", numero_funcionario);
-        getchar();
-        printf("\t==================================================\n\n");
-        printf("\t === Digite 0 para confirmar\n");
-        printf("\t=>");
-        scanf("%i", &voltar);
-        getchar();
-    }while(voltar);         
-}
-
-void tela_pesquisar_funcionario(){
-    char cpf_funcionario[14];
-    do{
-        system ("cls||clear");
-        printf("\t=========================================================\n");
-        printf("\t===============   Pesquisar Funcionarios   ==============\n");
-        printf("\t=========================================================\n\n");
-        printf("\t === Digite o CPF:    ");
-        scanf("%[0-9 .-/]", cpf_funcionario);
-        printf("\t==================================================\n\n");
-        printf("\t === Digite 0 para voltar\n");
-        printf("\t=>");
-        scanf("%i", &voltar);
-        getchar();
-    }while(voltar);
-}
-
-void tela_alterar_funcionario(){
-    char cpf_funcionario[14];
-    do{
-        system ("cls||clear");
-        printf("\t=========================================================\n");
-        printf("\t================   Alterar Funcionarios   ===============\n");
-        printf("\t=========================================================\n\n");
-        printf("\t === Digite o CPF:    ");
-        scanf("%[0-9 .-/]", cpf_funcionario);
-        printf("\t==================================================\n\n");
-        printf("\t === Digite 0 para confirmar\n");
-        printf("\t=>");
-        scanf("%i", &voltar);
-        getchar();
-    }while(voltar);
-}
-
-void tela_excluir_funcionario(){
-    char cpf_funcionario[14];
-    do{
-        system ("cls||clear");
-        printf("\t=========================================================\n");
-        printf("\t================   Excluir Funcionarios   ===============\n");
-        printf("\t=========================================================\n\n");
-        printf("\t === Digite o CPF:    ");
-        scanf("%[0-9 .-/]", cpf_funcionario);
-        printf("\t==================================================\n\n");
-        printf("\t === Digite 0 para confirmar a exclusao\n");
-        printf("\t=>");
-        scanf("%i", &voltar);
-        getchar();
-    }while(voltar);
 }
 
 void tela_compra(){
