+++ funcoes_funcionarios.h
@@ -1,0 +1,4 @@
+void tela_cadastrar_funcionario();
+void tela_pesquisar_funcionario();
+void tela_alterar_funcionario();
+void tela_excluir_funcionario();
