+++ funcoes_paciente.h
@@ -1,0 +1,4 @@
+void tela_cadastrar_paciente();
+void tela_pesquisar_paciente();
+void tela_alterar_paciente();
+void tela_excluir_paciente();
