+++ funcoes_servico.h
@@ -1,0 +1,4 @@
+void tela_cadastrar_servico();
+void tela_pesquisar_servico();
+void tela_alterar_servico();
+void tela_excluir_servico();
